[++] added unit tests [==] restructured code base
